+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 * Update eager loading ([#876](https://gitlab.com/yale-sdmp/vesta/issues/876)).
+* Fix inactive draw_memberships from interfering with special group formation ([#941](https://gitlab.com/yale-sdmp/vesta/issues/941)).
 
 ### Changed
 * Rename group size locking to restrict size locking ([811](https://gitlab.com/yale-sdmp/vesta/issues/811)).